Explicit TypeVisitable implementations
